+++ ocrd_segment/cli.py
@@ -10,6 +10,8 @@ from .replace_page import ReplacePage
 from .extract_pages import ExtractPages
 from .extract_regions import ExtractRegions
 from .extract_lines import ExtractLines
+from .extract_words import ExtractWords
+from .extract_glyphs import ExtractGlyphs
 from .extract_address import ExtractAddress
 from .extract_formdata import ExtractFormData
 from .classify_address_text import ClassifyAddressText
@@ -66,6 +68,16 @@ def ocrd_segment_extract_lines(*args, **kwargs):
 
 @click.command()
 @ocrd_cli_options
+def ocrd_segment_extract_words(*args, **kwargs):
+    return ocrd_cli_wrap_processor(ExtractWords, *args, **kwargs)
+
+@click.command()
+@ocrd_cli_options
+def ocrd_segment_extract_glyphs(*args, **kwargs):
+    return ocrd_cli_wrap_processor(ExtractGlyphs, *args, **kwargs)
+
+@click.command()
+@ocrd_cli_options
 def ocrd_segment_extract_address(*args, **kwargs):
     return ocrd_cli_wrap_processor(ExtractAddress, *args, **kwargs)
 
@@ -103,3 +115,4 @@ def ocrd_segment_classify_formdata_dummy(*args, **kwargs):
 @ocrd_cli_options
 def ocrd_segment_postcorrect_formdata(*args, **kwargs):
     return ocrd_cli_wrap_processor(PostCorrectFormData, *args, **kwargs)
+